GENERALIZE VARIABLE TYPE

+++ dunereco/LowEUtils/LowEUtils.h
@@ -298,7 +298,7 @@ namespace lowe
             const std::vector<std::pair<std::string, std::vector<double>>> fAdjOpFlashMaxPELightMap; // Light map file and histogram name for light map attenuation
             const std::vector<std::pair<std::string, std::vector<double>>> fAdjOpFlashPELightMap;    // Light map file and histogram name for PE attenuation
             const std::string fFlashMatchBy;                   // Method to match flashes ("maximum" or "light_map")
-            const int fFlashMatchByPELightMapExponent;         // Exponent for PE weighting in light map flash matching
+            const float fFlashMatchByPELightMapExponent;         // Exponent for PE weighting in light map flash matching
             std::unique_ptr<producer::ProducerUtils> producer; // Pointer to the ProducerUtils instance
     };
 } // namespace lowe
